refactor: rename data_types/src/chunk.rs -> data_types/src/chunk_metadata.rs
