+++ data_types/src/lib.rs
@@ -10,7 +10,7 @@
     clippy::clone_on_ref_ptr
 )]
 
-pub mod chunk;
+pub mod chunk_metadata;
 pub mod consistent_hasher;
 mod database_name;
 pub use database_name::*;
