+++ read_buffer/src/chunk.rs
@@ -8,7 +8,7 @@ use parking_lot::RwLock;
 use snafu::{OptionExt, ResultExt, Snafu};
 
 use arrow::record_batch::RecordBatch;
-use data_types::{chunk::ChunkColumnSummary, partition_metadata::TableSummary};
+use data_types::{chunk_metadata::ChunkColumnSummary, partition_metadata::TableSummary};
 use internal_types::{schema::builder::Error as SchemaError, schema::Schema, selection::Selection};
 use observability_deps::tracing::info;
