+++ src/commands/database/partition.rs
@@ -1,5 +1,5 @@
 //! This module implements the `partition` CLI command
-use data_types::chunk::ChunkSummary;
+use data_types::chunk_metadata::ChunkSummary;
 use data_types::job::Operation;
 use generated_types::google::FieldViolation;
 use influxdb_iox_client::{
